construct Input shape from both Spec and Status
We were only attempting to set the fields on an Input shape by looking
for a corresponding field in the CRD's Spec struct. This makes sense in
a way, since constructing, for example, the CreateBookInput shape would
by nature only take fields from the Book's desired state (i.e. the
BookSpec) struct.

However when constructing Input shapes like that used for ReadOne
operations, we need to look in the CRD's Status fields as well as the
Spec fields.

For example, consider the API Gateway V2 API's GetApiInput shape:

```go
type GetApiInput struct {
	_ struct{} `type:"structure"`

	// ApiId is a required field
	ApiId *string `location:"uri" locationName:"apiId" type:"string" required:"true"`
}
```

The Go code we were outputting for the `newDescribeRequestPayload()`
method in that generated service controller's `pkg/resource/api/sdk.go`
file looked like this:

```go
func (rm *resourceManager) newDescribeRequestPayload(
	r *resource,
) (*svcsdk.GetApiInput, error) {
	res := &svcsdk.GetApiInput{}

	return res, nil
}
```

There is a `GetApiInput.ApiId` field that isn't being set because the
`ApiId` value is stored in the `API.Status.APIID` field, not the
`API.Spec.APIID` field.

With this patch, we now output the proper Go code to populate the Input
shape with values from both Spec and Status fields in the CR:

```go

// newDescribeRequestPayload returns SDK-specific struct for the HTTP request
// payload of the Describe API call for the resource
func (rm *resourceManager) newDescribeRequestPayload(
	r *resource,
) (*svcsdk.GetApiInput, error) {
	res := &svcsdk.GetApiInput{}

	res.SetApiId(*r.ko.Status.APIID)

	return res, nil
}
```

Issue #127